requirements.txt
aiogram==2.25.1
aiohttp==3.8.5
aiosignal==1.3.1
asgiref==3.7.2
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.2.0
colorama==0.4.6
comm==0.1.3
debugpy==1.6.7
decorator==5.1.1
distlib==0.3.7
et-xmlfile==1.1.0
executing==1.2.0
filelock==3.12.3
frozenlist==1.4.0
google-auth==2.21.0
google-auth-oauthlib==1.0.0
gspread==5.10.0
icrawler==0.6.7
idna==3.4
ipykernel==6.25.0
ipython==8.14.0
jedi==0.18.2
jupyter_client==8.3.0
jupyter_core==5.3.1
lxml==4.9.3
magic-filter==1.0.11
matplotlib-inline==0.1.6
multidict==6.0.4
nest-asyncio==1.5.6
numpy==1.25.1
oauthlib==3.2.2
openpyxl==3.1.2
packaging==23.1
pandas==2.0.3
parso==0.8.3
pickleshare==0.7.5
platformdirs==3.9.1
psutil==5.9.5
psycopg==3.1.10
psycopg2-binary==2.9.7
pure-eval==0.2.2
py4j==0.10.9.7
pyasn1==0.5.0
pyasn1-modules==0.3.0
Pygments==2.15.1
pyTelegramBotAPI==4.12.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0.1
pyzmq==25.1.0
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
schedule==1.2.0
soupsieve==2.4.1
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16
virtualenv==20.24.4
wcwidth==0.2.6
XlsxWriter==3.1.6
yarl==1.9.2
